Pre Release Version
Added favicon svg.
Added a last updated p tag to index.html.
Added rating, ratingCount,  coupon, and primePrice variables to be displayed to the user.
Added helper function to help format the inputs of the price,  discount, and coupon.
Added to the server.js an svg contentType.
Added a backup List file.
Added more data to the list.json
Added to the Client side main.js file error handling for certain product values in the product list.
Updated the CSS to give certain labels color.
Added Animation of color called prime to the label with the prime price.